Fixed the issue with corrupt image display on multi threads
the issue was that memory of tmp matrix in current thread was cleaned up as soon
as the variable falls out of scope. this memory was used by the QImage to display the image
and resulted in garbage display after the memory was cleared.

The solution uses QMutex and QWaitCondition to synchronize the threads and ensure the variable stays
in scope until the window has finished rendering.

fixes #96